Mulitplexer message format changes for easier parsing

+++ LOOM_library/loom_mb1232.h
@@ -107,7 +107,7 @@ void package_mb1232(OSCBundle *bndl, char packet_header_string[])
 	char address_string[255];
 
 	sprintf(addressString, "%s%s", packet_header_string, "/mb1232_range");
-	bndl->add(addressString).add(state_mb1232.range);
+	bndl->add(addressString).add((int32_t)state_mb1232.range);
 }
